Rename rust files
