+++ lib/modules/manager/api.ts
@@ -83,7 +83,7 @@ import * as puppet from './puppet';
 import * as pyenv from './pyenv';
 import * as rubyVersion from './ruby-version';
 import * as runtimeVersion from './runtime-version';
-import * as rust from './rust';
+import * as rustToolchain from './rust-toolchain';
 import * as sbt from './sbt';
 import * as scalafmt from './scalafmt';
 import * as setupCfg from './setup-cfg';
@@ -189,7 +189,7 @@ api.set('puppet', puppet);
 api.set('pyenv', pyenv);
 api.set('ruby-version', rubyVersion);
 api.set('runtime-version', runtimeVersion);
-api.set('rust-toolchain', rust);
+api.set('rust-toolchain', rustToolchain);
 api.set('sbt', sbt);
 api.set('scalafmt', scalafmt);
 api.set('setup-cfg', setupCfg);
