+++ lib/modules/manager/rust-toolchain/extract.ts
